Make changes in props and code enhancement
